+++ packages/client/src/client.ts
@@ -8,6 +8,7 @@ import {
 import { Binary, HexString } from "@polkadot-api/substrate-bindings"
 import {
   SubstrateClient,
+  UnsubscribeFn,
   createClient as createRawClient,
 } from "@polkadot-api/substrate-client"
 import { toHex } from "@polkadot-api/utils"
@@ -260,6 +261,43 @@ export function createClient(
     params: Params,
   ) => Promise<Reply> = rawClient.request
 
+  const _subscribe: PolkadotClient["_subscribe"] = (
+    method,
+    unsubscribeMethod,
+    params,
+  ) =>
+    new Observable((obs) => {
+      let unsubInner: UnsubscribeFn | null = null
+      let subId: string | null = null
+
+      const sendUnsubscribe = () => {
+        // Fire-and-forget
+        subId != null && rawClient.request(unsubscribeMethod, [subId])
+      }
+
+      rawClient._request(method, params, {
+        onSuccess: (res: string, follow) => {
+          subId = res
+          if (obs.closed) {
+            sendUnsubscribe()
+            return
+          }
+          unsubInner = follow(subId, {
+            next: (data: any) => obs.next(data),
+            error: (e) => obs.error(e),
+          })
+        },
+        onError(e) {
+          obs.error(e)
+        },
+      })
+
+      return () => {
+        sendUnsubscribe()
+        unsubInner?.()
+      }
+    })
+
   const { broadcastTx$ } = client
 
   const getMetadata$ = (at: HexString) =>
@@ -322,6 +360,7 @@ export function createClient(
     },
 
     _request,
+    _subscribe,
   }
 
   ;(result as any).___INTERNAL_DO_NOT_USE = chainHead